Define own GetUtxosResponse type
