+++ rs/bitcoin/ckbtc/minter/src/test_fixtures.rs
@@ -1,9 +1,9 @@
 use crate::lifecycle::init::{BtcNetwork, InitArgs};
 use crate::Timestamp;
-use crate::{lifecycle, ECDSAPublicKey};
+use crate::{lifecycle, ECDSAPublicKey, GetUtxosResponse};
 use candid::Principal;
 use ic_base_types::CanisterId;
-use ic_btc_interface::{GetUtxosResponse, OutPoint, Utxo};
+use ic_btc_interface::{OutPoint, Utxo};
 use icrc_ledger_types::icrc1::account::Account;
 use std::time::Duration;
 
@@ -105,10 +105,6 @@ pub fn quarantined_utxo() -> Utxo {
 pub fn get_uxos_response() -> GetUtxosResponse {
     GetUtxosResponse {
         utxos: vec![],
-        tip_block_hash: hex::decode(
-            "00000000000000000002716d23b6b02097a297a84da484c7a9b6427a999112d8",
-        )
-        .unwrap(),
         tip_height: 871160,
         next_page: None,
     }
@@ -117,11 +113,11 @@ pub fn get_uxos_response() -> GetUtxosResponse {
 pub mod mock {
     use crate::management::CallError;
     use crate::updates::update_balance::UpdateBalanceError;
-    use crate::{CanisterRuntime, GetUtxosRequest};
+    use crate::{CanisterRuntime, GetUtxosRequest, GetUtxosResponse};
     use async_trait::async_trait;
     use candid::Principal;
     use ic_btc_checker::CheckTransactionResponse;
-    use ic_btc_interface::{GetUtxosResponse, Utxo};
+    use ic_btc_interface::Utxo;
     use icrc_ledger_types::icrc1::account::Account;
     use icrc_ledger_types::icrc1::transfer::Memo;
     use mockall::mock;
